Added custom script functionality (#59)
* Added custom script functionality


* Running pretest always when defined

+++ upstream/test/test.sh
@@ -31,6 +31,7 @@ OP_TEST_EXEC_EXTRA=${OP_TEST_EXEC_EXTRA-"-e container_tool=podman"}
 OP_TEST_RUN_MODE=${OP_TEST_RUN_MODE-"privileged"}
 OP_TEST_LABELS=${OP_TEST_LABELS-""}
 OP_TEST_PROD=${OP_TEST_PROD-0}
+OP_TEST_PRETEST_CUSTOM_SCRIPT=${OP_TEST_PRETEST_CUSTOM_SCRIPT-""}
 OP_TEST_DEBUG=${OP_TEST_DEBUG-0}
 OP_TEST_DRY_RUN=${OP_TEST_DRY_RUN-0}
 OP_TEST_FORCE_INSTALL=${OP_TEST_FORCE_INSTALL-0}
@@ -429,7 +430,13 @@ for t in $TESTS;do
         echo -e "[$t] Reseting kind cluster ..."
         run $DRY_RUN_CMD ansible-pull -U $OP_TEST_ANSIBLE_PULL_REPO -C $OP_TEST_ANSIBLE_PULL_BRANCH $OP_TEST_ANSIBLE_DEFAULT_ARGS --tags reset
     fi
-
+    if [ -n "$OP_TEST_PRETEST_CUSTOM_SCRIPT" ];then
+        echo "Running custom script '$OP_TEST_PRETEST_CUSTOM_SCRIPT' ..."
+        [ -f $OP_TEST_PRETEST_CUSTOM_SCRIPT ] || { echo "Custom script '$OP_TEST_PRETEST_CUSTOM_SCRIPT' was not found. Exiting ..."; exit 1; }
+        [[ -x "$OP_TEST_PRETEST_CUSTOM_SCRIPT" ]] || { echo "Custom script '$OP_TEST_PRETEST_CUSTOM_SCRIPT' is not executable. Do 'chmod +x $OP_TEST_PRETEST_CUSTOM_SCRIPT' first !!! Exiting ..."; exit 1; }
+        run $OP_TEST_PRETEST_CUSTOM_SCRIPT
+        echo "Custom script '$OP_TEST_PRETEST_CUSTOM_SCRIPT' done ..."
+    fi
     echo -e "[$t] Running test ..."
     [[ $OP_TEST_DEBUG -ge 3 ]] && echo "OP_TEST_EXEC_EXTRA=$OP_TEST_EXEC_EXTRA"
     $DRY_RUN_CMD $OP_TEST_CONTAINER_TOOL rm -f $OP_TEST_NAME > /dev/null 2>&1
